+++ src/test/java/kitchenpos/order/domain/OrderTest.java
@@ -1,0 +1,13 @@
+package kitchenpos.order.domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+public class OrderTest {
+
+    @DisplayName("주문 생성")
+    @Test
+    void create() {
+
+    }
+}
